requirements.txt
pandas
numpy
matplotlib
seaborn
jupyter
pytest
tqdm
telethon
python-dotenv
dbt-postgres
psycopg2
ultralytics
sqlalchemy